+++ test/e2e/btc_rewards_distribution_e2e_test.go
@@ -37,7 +37,7 @@ const (
 	wDel2             = "del2"
 	wFp1              = "fp1"
 	wFp2              = "fp2"
-	numPubRand        = uint64(350)
+	numPubRand        = uint64(600)
 )
 
 type BaseBtcRewardsDistribution struct {